beta 7.6.3.3 source only
SongbookUpdater
=====
implemented algorithm for parsing results

+++ src/stone/modules/SongbookUpdater.java
@@ -208,7 +208,9 @@ public final class SongbookUpdater implements Module {
 		}
 		io.close(in);
 		container.fill();
-
+		if (master.isInterrupted())
+			return;
+		
 		final File masterPluginData = songbookPlugindataPath.toFile();
 		masterPluginData.deleteOnExit();
